refactor: make some of the extensions package public
Moving the loading and validation to the machinery package, so that we
can import and use that from other projects.
