+++ pkg/machinery/extensions/testdata/bad/badpaths/manifest.yaml
