+++ pkg/machinery/extensions/testdata/bad/emptymanifest/manifest.yaml
