+++ pkg/machinery/extensions/testdata/bad/norootfs/manifest.yaml
