+++ pkg/machinery/extensions/testdata/good/extension1/rootfs/usr/local/lib/a.so.1
@@ -1,0 +1,1 @@
+a.so
